Removed level managers
