+++ changelogs/2025/2025_01_20 - Game Play.md
@@ -1,4 +1,4 @@
-# FUN Command Line Game Log: 11.27.2023 - 
+# FUN Command Line Game Log: 11.27.2023 -
 <!-- Update: Current Log date -->
 
 Last Updated: 1.20.225 <!-- Update with previous log date -->
@@ -13,11 +13,13 @@ Version: 0.7.x
  -->
 
 ## MOTD - 1.20.2025
+
 -------------------
 
 Refreshing and learning code for udpates and development
 
-Continue working on game init process with 
+Continue working on game init process with
+
 - Create npcs and npc manager for level difficulty testing process
 - level generation and difficulty process
 - Define new roles, equipment, and equipment type
@@ -34,6 +36,7 @@ Need to create level pack manager
 - [ ] [#65](https://github.com/jevinevans/Game/issues/65)
 
 ### Create Issues
+
 - [ ] Level Pack creation [12.7.2023]
 - [ ] Using the NPCS csv in docs, create new roles, equipment type, and new equipment to accomediate the NPCs
 - [ ] Update Char_Manager to list all roles and not ask the user for the role type they want. Select role use that armor type as the char armor type [12.7.2023]
@@ -41,9 +44,7 @@ Need to create level pack manager
 - [ ] Create a manager class and then use it as a base class for the individual managers. It should simply management, and will allow for proper calls instead of loading on run
 - [ ] Equipment Manager, create break down when view armor to view item type only (i.e., view all heads, view all chest)
 
-
 ## Changes
-------------------------------
 
 ### Added
 
@@ -54,16 +55,13 @@ Need to create level pack manager
 - Added Nonplayablecharater DB prefix
 
 ### Fixed
+
 - Fixed issued with number validation default value
 
 ### Informational
 
+## Long Term Priority
 
-------------------------------
-
-
-# Long Term Priority
----------------------------
 - Game Format
   - level creation
   - Combat initiation
@@ -81,4 +79,3 @@ Need to create level pack manager
 
 - build a npc manager
   - Create a default list of enemies that should be used
-
