+++ funclg/managers/__init__.py
@@ -8,8 +8,6 @@ from . import abilities_manager as ability_man
 from . import character_manager as char_man
 from . import equipment_manager as equip_man
 from . import game_manager as game_man
-from . import level_manager as level_man
-from . import level_pack_manager as lvl_pack_man
 from . import npc_manager as npc_man
 from . import roles_manager as role_man
 
@@ -19,11 +17,10 @@ CHAR_MANAGERS = [
     ability_man,
     equip_man,
 ]
-GAME_MANAGERS = [game_man, level_man, lvl_pack_man, npc_man]
+GAME_MANAGERS = [game_man, npc_man]
 
 ALL_MANAGERS = CHAR_MANAGERS + GAME_MANAGERS
 
-GAME_MANAGERS = [game_man, level_man]
 
 
 def build_manager_menu():
@@ -54,12 +51,10 @@ def build_game_menu():
 
 
 def load_data():
-    # for manager in ALL_MANAGERS:
-    for manager in CHAR_MANAGERS:
+    for manager in ALL_MANAGERS:
         manager.load_data()
 
 
 def save_data():
-    # for manager in ALL_MANAGERS:
-    for manager in CHAR_MANAGERS:
+    for manager in ALL_MANAGERS:
         manager.export_data()
